+++ ek-tm4c1294xl/project3/serial.h
@@ -60,28 +60,28 @@
 // Prototypes for the Serial interface functions.
 //
 //*****************************************************************************
-extern tBoolean SerialSendFull(unsigned long ulPort);
-extern void SerialSend(unsigned long ulPort, uint8_t ucChar);
-extern long SerialReceive(unsigned long ulPort);
-extern unsigned long SerialReceiveAvailable(unsigned long ulPort);
-extern void SerialSetBaudRate(unsigned long ulPort, unsigned long ulBaudRate);
-extern unsigned long SerialGetBaudRate(unsigned long ulPort);
-extern void SerialSetDataSize(unsigned long ulPort, uint8_t ucDataSize);
-extern uint8_t SerialGetDataSize(unsigned long ulPort);
-extern void SerialSetParity(unsigned long ulPort, uint8_t ucParity);
-extern uint8_t SerialGetParity(unsigned long ulPort);
-extern void SerialSetStopBits(unsigned long ulPort, uint8_t ucStopBits);
-extern uint8_t SerialGetStopBits(unsigned long ulPort);
-extern void SerialSetFlowControl(unsigned long ulPort,
+extern tBoolean SerialSendFull(uint32_t ulPort);
+extern void SerialSend(uint32_t ulPort, uint8_t ucChar);
+extern long SerialReceive(uint32_t ulPort);
+extern uint32_t SerialReceiveAvailable(uint32_t ulPort);
+extern void SerialSetBaudRate(uint32_t ulPort, uint32_t ulBaudRate);
+extern uint32_t SerialGetBaudRate(uint32_t ulPort);
+extern void SerialSetDataSize(uint32_t ulPort, uint8_t ucDataSize);
+extern uint8_t SerialGetDataSize(uint32_t ulPort);
+extern void SerialSetParity(uint32_t ulPort, uint8_t ucParity);
+extern uint8_t SerialGetParity(uint32_t ulPort);
+extern void SerialSetStopBits(uint32_t ulPort, uint8_t ucStopBits);
+extern uint8_t SerialGetStopBits(uint32_t ulPort);
+extern void SerialSetFlowControl(uint32_t ulPort,
                                  uint8_t ucFlowControl);
-extern uint8_t SerialGetFlowControl(unsigned long ulPort);
-extern void SerialSetFlowOut(unsigned long ulPort, uint8_t ucFlowValue);
-extern uint8_t SerialGetFlowOut(unsigned long ulPort);
-extern void SerialPurgeData(unsigned long ulPort,
+extern uint8_t SerialGetFlowControl(uint32_t ulPort);
+extern void SerialSetFlowOut(uint32_t ulPort, uint8_t ucFlowValue);
+extern uint8_t SerialGetFlowOut(uint32_t ulPort);
+extern void SerialPurgeData(uint32_t ulPort,
                             uint8_t ucPurgeCommand);
-extern void SerialSetDefault(unsigned long ulPort);
-extern void SerialSetFactory(unsigned long ulPort);
-extern void SerialSetCurrent(unsigned long ulPort);
+extern void SerialSetDefault(uint32_t ulPort);
+extern void SerialSetFactory(uint32_t ulPort);
+extern void SerialSetCurrent(uint32_t ulPort);
 extern void SerialInit(void);
 
 #endif // __SERIAL_H__
